+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Position files now can include virtual components. (#106)
 - Support for variants on KiCost output. (#106)
 - `--cli-order` option to generate outputs in arbitrary order. (#106)
+- QR codes generation: symbols and footprints. (#93)
 
 ### Changed
 - Internal BoM: now components with different Tolerance, Voltage, Current